Tags now show up also on the first time "tags" is selected
Now fetches tags and keywords twice when page is completed instead of
once every time the button is pressed. No idea why it works on the
second round but not the first.

Also added .desktop and a launcer icon. Icon does not show up.

Changes in .pro and main.cpp due to lack of qmlapplicationviewer.
Compiles again.